// angular-frontend/src/app/generated/model/tagTimeRegistrationRequest.ts
/**
 * TriReg2 User API
 * An OpenAPI specification for time registrations to be used by a user
 *
 * The version of the OpenAPI document: 0.2.0
 * Contact: example@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TagTimeRegistrationRequest { 
    timeRegistrationId: number;
    tagConfigurationId: number;
    value?: string;
}